requirements.txt
flask>=3.0.0
google-genai>=1.0.0
opencv-python>=4.9.0
numpy>=1.26.0
pandas>=2.2.0
Werkzeug>=3.0.0
python-dotenv>=1.0.0
